fix(F026): Address code review findings
1. get_pending_requirements() now filters by event_id
   - Queries EventAssemblyTarget for assembly compositions
   - Queries EventRecipientPackage for package compositions
   - Uses OR to include compositions from either path

2. get_event_packaging_needs() uses float instead of int for quantity
   - Fixes truncation of fractional packaging quantities
   - Ensures accurate estimated cost calculation

Fixes issues identified in Cursor code review.

+++ src/services/event_service.py
@@ -1521,7 +1521,7 @@ def get_event_packaging_needs(event_id: int) -> Dict[str, PackagingNeed]:
                 # Get estimated cost from packaging service
                 try:
                     estimated_cost = packaging_service.get_estimated_cost(
-                        product_name, int(total_needed), session=session
+                        product_name, float(total_needed), session=session
                     )
                 except Exception:
                     estimated_cost = None
